Added plotting options; updated get_obs
The user can now select individual plot/analysis options from a menu, and
the do_analysis method retains (some of) the data in the Beans object for
multiple calls.

In parallel investigated the issue with varying return parameter lists in
get_obs; this is necessary because of the additional information required
when gti_checking is on. As with the other routines, removed the
gti_checking parameter from the input list, now is indicated purely by the
gtiname param (set to None to turn gti_checking off)