General cleanup, simplify data reading during init
Instead of passing the individual parameters to get_obs, we now just pass
the Beans object, from which the necessary parameters are extracted (and
to which the necessary data is added). This change simplifies the calling
procedure (and removes the need for a separate initialise.py function).

Simplified the assembly of the inverse sigma values in the likelihood
calculation

Removed the various globals, which were no longer used

Removed the x attribute, which is not used; in the various calls just pass
None