Added fig6, fig8 plots; simplify posterior stats
Added options to reproduce Figures 6 and 8 from Goodwin et al. 2019.
Figure 8 requires model predictions for xi_b, xi_p from the concord
package; if that's not available a warning will be shown, but the plot
will still be produced (omitting those lines).

Simplified the parameter calculations and the way the predictions are
extracted, and also applied the burnin parameter also to the blobs