Reduced redundancy in blobs, model dict
Adjusted the generate_burst_train and burstensemble routines to include a
full_model option, which (if set to False, as is the default) omits the
redundant (input parameters) being included in the dict returned.

This reduction in the size of the dict is then passed through to the .h5
file, reducing the size of each blob (and the overall .h5 file) by an
estimated ~40%.

The do_analysis routine has been tweaked to apply the burnin also to the
blobs, and extract the parameters more efficiently.