Added spline interpolation, simplified function calls
Modifed runmodel (and the other routines that it calls) to pass a Beans
object instead of individual parameters, following the earlier changes to
get_obs

Moved get_a_b and mean_flux to the main beans.py file, renaming mean_flux
to mean_flux_linear. Added a spline interpolation option with a
user-selectable smoothing factor (defaulting to 0.02) via the
mean_flux_spline function, the specific choice of averaging function is
now stored as part of the Beans object

Minor updates to the config file reading/writing, tests etc. to
accommodate the above changes