Parameters dist, xi_b, xi_p replace ratios
Modified much of the code to now work directly with the distance (in kpc),
and the burst and persistent flux anisotropies, rather than the ratio
values.

Wrote a new functions, flux_to_mdot to manage the conversion there;
checked against concord, seems OK, ditto for the fluence conversion. Also
defined a couple of constants which are used in the conversion.

Modified the code to work without systematic errors, which can be
triggered by passing an 8-element parameter tuple on init. Not extensively
tested as yet.