Added corr function; fixed fluen percentile bug
Can now specify a function to modify the predicted burst parameters from
Settle. An example is provided, corr_goodwin19, replicating the conditions
in the simulation of Goodwin et al. (2019)

corr file option is now stored in the config file, and checked against the
input value when you read the values.

Also fixed a bug in do_analysis where the percentiles for the fluences
were incorrectly calculated.