Now handles inhomogeneous model results
Had issues before when the times &c arrays extracted from the blobs had
different numbers of bursts. This can happen when you're exploring
different areas of parameter space with different overall burst rates. Now
traps this and identifies the two (or more) classes of model, for the
comparison plot; and also translates this into the samples for the corner
plots.