Added grid interpolation for burstensemble's
Preliminary version with grid interpolation via the multiepoch_mcmc
routine, for which it's necessary to specify the path (currently
hardwired). User can specify model='grid_interp' at __init__ to choose this
functionality.
Added the routines grid_interp.py and the accompanying prior,
prior_grid which must be used.
Currently only works with the burstensemble mode; bit more tricky to do it
with the train mode, but that's likely not useful at the moment with the
present grid, which covers a narrow range of mdot (assembled for use with
GS 1826-24; see Johnston et al., 2020, https://doi.org/10.1093/mnras/staa1054)
Model is now stored with the bean object, and added some kwargs variously
so that all the routines can handle all the options (even if some are
ignored for specific models).
Choice of model is written to the config file; older config files will be
assumed to be using settle.
mdot still needs to be specified separately; currently hardwired to the
grid version (no dependence on X, g).