Now includes likelihood contribution from peak fluxes
Fixed the likelihood calculation in lnlike, lnlike_sys, which was missing
a factor of 2/pi.

Modifed get_data.py to allow reading in burst data files that also have a
peak flux, error, and PRE flag column. These data are now used in lnlike,
lnlike_sys to add a contribution to the likelihood, for the moment only
for non-PRE bursts (but can easily be extended to PRE bursts).